Rewrite StatusHeader as FC with TS
This is not properly typed yet.

Also slight refactor to combine wrapped functions into one. This still works.
