+++ app/javascript/flavours/polyam/components/status.jsx
@@ -26,7 +26,7 @@ import AttachmentList from './attachment_list';
 import { getHashtagBarForStatus } from './hashtag_bar';
 import StatusActionBar from './status_action_bar';
 import StatusContent from './status_content';
-import StatusHeader from './status_header';
+import { StatusHeader } from './status_header';
 import StatusIcons from './status_icons';
 import StatusPrepend from './status_prepend';
 import StatusReactions from './status_reactions';
